Cookie jar now sits on the RestContext
The oae-rest module no longer takes care of keeping cookie jars around.
These are stored on the RestContext object and it's up to the client
to ensure that these don't time out.

This also allows for the PP to fetch a new session for each content object.

I've also added an error listener in the PP on the oae rest module