Added an autoOrient utility in oae-util/image
Rather then adding autoOrient on all of the ImageUtil methods, I've added
an extra method that can be used for this. `autoOrient` simply brings
to much overhead and a performance degradation to execute it on each
image operation.

For example, doing an `autoOrient` once on the source preview image rather
then on each single one, results in a performance improvement of around
36 percent.

I've also added `autoOrient` to the profile pictures so that they can be
displayed correctly in the browser.