#671 - Allow for redirecting with the external auth strategies
Specifying a `redirectUrl` request parameter in the external
authentication strategies will redirect the user to that URL once they
return to the application. If no value is provided, `/me` will be used.

I also took the opportunity to streamline the external auth strategies
a little bit so we don't need to paste the complex passport boilerplate
in each strategy's REST file.

Currently supported strategies:
 * CAS
 * Facebook
 * Google
 * Shibboleth
 * Signed authentication
 * Twitter