787 - Comment creation responses should include profile picture URLs
This regression was caused because we were using the PrincipalsDAO rather
than the PrincipalsUtil. The Util is responsible for adding signed profile
picture URLs.

It seems like comments and discussion messages are the only places
that are currently taking the user object from the request context and
returning it in the response.