#819 - Using `previewsId` for scoping PDF CSS file
This commit scopes the CSS rules in the combined.css file for PDF
previews to the `previewsId` rather than the `revisionId`. The benefit
of using the `previewsId` is that we can reuse the previews when
restoring an earlier revision.

The CSS class name that was used to scope pdf2html's CSS rules is stored
in the `previews` object on the content profile as a `cssScopeClass`
property.