Added a `resetAggregationForActivityStreams` to the activity aggregator
This allows you to halt activity aggregation for specific streams.

Each time an aggregate key for an activity stream is updated, we add the key in a sorted set of "active aggregate keys" for that stream.
Each time we insert a key in the set, we give it a rank of the current millis since epoch.
This allows us to easily remove aggregate keys that have expired.
The set itself also expires (maxExpiry seconds after the last time a key was inserted into it)