Fix for intermittent push unit test failures
Bug: The activties were being delivered on the websocket before we had
bound the `message` event listener.

Fix: Bind the event listener in the same process tick as we perform the
action for which we should receive a message on the socket.

The tests have also been cleaned up a bit as they were performing
unnecessary activity collections (push notification happens on routing,
there’s no need for aggregation/collection)