Count `log().error` calls with the Telemetry API
The log().error function is now wrapped so the number of times it gets
called (and from where) can be tracked through the Telemetry API. This
will allow us to put Nagios checks in place that measure error rates.

In the future statsd hooks could be put in place as well.