Handle properly the creation of loodle of a loodle content with the mechanism of folder.
When a loodle is created inside a folder, the users of the folder have
a « readonly » access to the loodle.
When a loodle is added to a folder, the users of the folder gain a
« readonly » access to the loodle.
When a loodle is removed from a folder, the users of the folder lose
their « readonly » access to the loodle.