Add functional tests "create a loodle in many folders".
Check if it creates correctly the loodle.
Check if the loodle is accessible in readonly mode to the members of
the folders.
Check if the loodle is accessible to the creator of the users
explicitly invited.
Check if the only participants displayed in the loodle data are the
users explicitly invited and the creator.