Add functional tests "Create a loodle in a group".
Check if the loodle is correctly created.
Check if the loodle is accessible by the group members.
Check if the loodle is accessible by the users explicitly invated or
the creator.
Check if the participants displayed in the loodle data are the
users explicitly invited, the creator and the group members.