Add functional tests "create a loodle in many groups". Correction of the loodle api to avoid a problem of "insufficient privilege" when creating a loodle.
Check if the loodle is correctly created.
Check if the loodle is accessible by the groups members.
Check if the loodle is accessible by the users explicitly invited and
the creator.
Check if the participants displayed in the loodle data are the
users explicitly invited, the creator and the groups members.
Check that any user who have multiple access to this loodle (he/she is
a member of two groups which have access to the loodle for example) is
only added one time as a participant.
Check that the case where a group is a member of another group in which
a loodle is created is handle properly : every user of the groups are
added to the loodle.