Add functional tests "create a loodle in folders and groups".
Check if the loodle is correctly created.
Check if the loodle is accessible by the folder(s) members in read only
mode.
Check if the loodle is accessible by the groups members.
Check if the loodle is accessible by the users explicitly invited and
the creator.
Check if the participants displayed in the loodle data are the
users explicitly invited, the creator and the groups members.
Check that any user who have multiple access to this loodle (he/she is
a member of two groups which have access to the loodle for example) is
only added one time as a participant.
Check that the case where a group is a member of another group in which
a loodle is created is handle properly : every user of the groups are
added to the loodle.