Add functional tests "add a loodle to a folder". Correction of the loodle api according to these tests and refactoring of the code.
Refactoring of the code added in « oae-content/lib/api.js » :
suppression of most of the added code for the loodle functionality. Add
an event when an user get a content : « GET_CONTENT ».
Suppression of two useless events (created some time ago for the loodle
functionality) : « EDITED_LOODLE_SCHEDULES » and
« EDITED_LOODLE_VOTES ». Creation of the event « GET_CONTENT ».
Modification of the way the event ADDED_CONTENT_ITEMS is emitted : a
callback function is now being called when every listener to this event
has done its work.
Refactoring of the oae-loodle/lib.api.js file to detect some events
emitted in other modules and do the appropriate works instead of adding
some codes in those modules.