Refactoring : suppression of the code handling the loodle notifications from the oae-content module to move it to the oae-loodle module.
Creation of the oae-loodle/lib/constants.js file that list all the
events and activities bind to a loodle.
Creation of the oae-loodle/lib/activity.js file that listen and
register the differents events and activities bind to a loodle.
Update the oae-loodle/lib/init.js file that take into account these
activities.
Suppression of the corresponding useless portions of code of the
oae-content module.